+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1507,7 +1507,7 @@ class Evaluations4(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-07-31T10:20:41.310Z")
+    ] = parse_datetime("2026-07-31T10:42:15.495Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1652,7 +1652,7 @@ class Evaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-07-31T10:20:41.309Z")
+    ] = parse_datetime("2026-07-31T10:42:15.495Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1797,7 +1797,7 @@ class Evaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-07-31T10:20:41.309Z")
+    ] = parse_datetime("2026-07-31T10:42:15.494Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1942,7 +1942,7 @@ class Evaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-07-31T10:20:41.308Z")
+    ] = parse_datetime("2026-07-31T10:42:15.494Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -2053,7 +2053,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-31T10:20:27.912Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-31T10:42:00.647Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")