+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -40,9 +40,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-31T10:20:29.856Z"
+    created: Optional[str] = "2026-07-31T10:42:02.740Z"
 
-    updated: Optional[str] = "2026-07-31T10:20:29.856Z"
+    updated: Optional[str] = "2026-07-31T10:42:02.740Z"
 
     updated_by_id: OptionalNullable[str] = UNSET