+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYVV1FCYMA9PRYT1YZ9JBYWQ"
+        "01KYVW8Z835ZNTFR085N6VFR1V"
     )
     r"""The unique identifier of the data source"""