+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYVV1FD1YM9NWEY2CHBGXB2A"
+        "01KYVW8Z86MXBNNAQKZ3K771N9"
     )
     r"""The unique identifier of the data source"""