+++ packages/orq-rc/RELEASES.md
@@ -9899,3 +9899,13 @@ Based on:
 - [python v4.13.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.23 - packages/orq-rc
+
+## 2026-07-31 10:45:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.24 - packages/orq-rc